Refactor of Matrix elements to make it reusable
